Add mesa drivers

+++ data/modify-driver.sh
@@ -37,12 +37,22 @@ fi
 
 if dpkg -s "$1"
 then
-	pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt remove $pkg -y && sudo apt autoremove -y"
+	if echo $pkg | grep -i mesa
+	then
+		zenity --error --text "the following driver "$1" can not be removed only swapped"
+	else
+		pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt remove $pkg -y && sudo apt autoremove -y"
+	fi
 else
 	if echo $pkg | grep -i nvidia
  	then
 		pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt update -y && apt purge nvidia-driver-* -y && apt install $pkg -y && sudo apt autoremove -y"
 	else
-		pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt update -y && apt install $pkg -y && sudo apt autoremove -y"
+		if echo $pkg | grep -i mesa-hybrid
+		then
+			pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt update -y && apt install mesa-stable -y && apt install mesa-hybrid -y && sudo apt autoremove -y"
+		else
+			pkexec env DISPLAY=$DISPLAY XAUTHORITY=$XAUTHORITY bash -c "apt update -y && apt install $pkg -y && sudo apt autoremove -y"
+		fi
 	fi
 fi
